moved modules by bam
