+++ src/blocks/popup/__image/PopupImage.js
@@ -1,4 +1,4 @@
-import Popup from './Popup.js';
+import Popup from '../Popup.js';
 export default class PopupImage extends Popup {
     constructor(paramPopup) {
         super(paramPopup);
